+++ lib/src/main/java/com/mastercard/gateway/ats/domain/CardServiceRequest.kt
@@ -1,11 +1,7 @@
 package com.mastercard.gateway.ats.domain
 
 import org.simpleframework.xml.*
-
-import java.math.BigInteger
 import java.util.*
-
- import java.util.Date
 
 @Root(name = "CardServiceRequest", strict = false)
 class CardServiceRequest : ATSMessage {
@@ -15,27 +11,27 @@ class CardServiceRequest : ATSMessage {
     var loyalty: CardServiceRequest.Loyalty? = null
     @field:Element(name = "CardCircuitCollection", required = false)
     var cardCircuitCollection: CardServiceRequest.CardCircuitCollection? = null
-    @field:Element(name = "OriginalTransaction", required=false)
+    @field:Element(name = "OriginalTransaction", required = false)
     var originalTransaction: CardServiceRequest.OriginalTransaction? = null
-    @field:Element(name = "TotalAmount", required=false)
+    @field:Element(name = "TotalAmount", required = false)
     var totalAmount: TotalAmountType? = null
-    @field:Element(name = "SaleItem", required=false)
+    @field:Element(name = "SaleItem", required = false)
     var saleItem: List<SaleItemType>? = null
-    @field:Element(name = "CardValue", required=false)
+    @field:Element(name = "CardValue", required = false)
     var cardValue: CardValueType? = null
-    @field:Element(name = "EncryptedCardValue", required=false)
+    @field:Element(name = "EncryptedCardValue", required = false)
     var encryptedCardValue: EncryptedCardValueType? = null
-    @field:Element(name = "Acquirer", required=false)
+    @field:Element(name = "Acquirer", required = false)
     var acquirer: Acquirer? = null
-    @field:Element(name = "PrivateData", required=false)
+    @field:Element(name = "PrivateData", required = false)
     var privateData: List<Any>? = null
-    @field:Element(name = "Encryption", required=false)
+    @field:Element(name = "Encryption", required = false)
     var encryption: Any? = null
-    @field:Element(name = "Validation", required=false)
+    @field:Element(name = "Validation", required = false)
     var validation: Any? = null
-    @field:Attribute(name = "ApplicationSender", required=false)
+    @field:Attribute(name = "ApplicationSender", required = false)
     var applicationSender: String? = null
-    @field:Attribute(name = "POPID", required=false)
+    @field:Attribute(name = "POPID", required = false)
     var popid: String? = null
     @field:Attribute(name = "RequestType", required = true)
     lateinit var requestType: CardRequestType
@@ -43,24 +39,24 @@ class CardServiceRequest : ATSMessage {
     lateinit var workstationID: String
     @field:Attribute(name = "RequestID", required = true)
     lateinit var requestID: String
-    @field:Attribute(name = "ReferenceNumber", required=false)
+    @field:Attribute(name = "ReferenceNumber", required = false)
     var referenceNumber: String? = null
 
 
     class Acquirer {
-        @field:Attribute(name = "TerminalID", required=false)
+        @field:Attribute(name = "TerminalID", required = false)
         var terminalID: String? = null
-        @field:Attribute(name = "MerchantID", required=false)
+        @field:Attribute(name = "MerchantID", required = false)
         var merchantID: String? = null
-        @field:Attribute(name = "AcquirerID", required=false)
+        @field:Attribute(name = "AcquirerID", required = false)
         var acquirerID: String? = null
-        @field:Attribute(name = "MerchantReference", required=false)
+        @field:Attribute(name = "MerchantReference", required = false)
         var merchantReference: String? = null
-        @field:Attribute(name = "vTIDClient", required=false)
+        @field:Attribute(name = "vTIDClient", required = false)
         var vtidClient: String? = null
-        @field:Attribute(name = "vTIDPassword", required=false)
+        @field:Attribute(name = "vTIDPassword", required = false)
         var vtidPassword: String? = null
-        @field:Attribute(name = "CreditPlan", required=false)
+        @field:Attribute(name = "CreditPlan", required = false)
         var creditPlan: String? = null
     }
 
@@ -73,25 +69,25 @@ class CardServiceRequest : ATSMessage {
         class CardCircuit {
             @field:Text
             lateinit var value: String
-            @field:Attribute(name = "CardCircuitState", required=false)
+            @field:Attribute(name = "CardCircuitState", required = false)
             var cardCircuitState: CardCircuitStateType? = null
         }
     }
 
     class Loyalty {
 
-        @field:Element(name = "LoyaltyCard", required=false)
+        @field:Element(name = "LoyaltyCard", required = false)
         var loyaltyCard: CardServiceRequest.Loyalty.LoyaltyCard? = null
-        @field:Element(name = "MOPrule", required=false)
+        @field:Element(name = "MOPrule", required = false)
         var moPrule: CardServiceRequest.Loyalty.MOPrule? = null
-        @field:Element(name = "LoyaltyAmount", required=false)
+        @field:Element(name = "LoyaltyAmount", required = false)
         var loyaltyAmount: Float? = null
         @field:Attribute(name = "LoyaltyFlag", required = true)
         var loyaltyFlag: Boolean = false
 
 
         class LoyaltyCard : CardTrack() {
-            @field:Attribute(name = "LoyaltyPAN", required=false)
+            @field:Attribute(name = "LoyaltyPAN", required = false)
             var loyaltyPAN: String? = null
         }
 
@@ -104,54 +100,54 @@ class CardServiceRequest : ATSMessage {
     }
 
     class OriginalTransaction {
-        @field:Attribute(name = "TerminalID", required=false)
+        @field:Attribute(name = "TerminalID", required = false)
         var terminalID: String? = null
-        @field:Attribute(name = "TerminalBatch", required=false)
+        @field:Attribute(name = "TerminalBatch", required = false)
         var terminalBatch: String? = null
-        @field:Attribute(name = "STAN", required=false)
+        @field:Attribute(name = "STAN", required = false)
         var stan: Long? = null
-        @field:Attribute(name = "TimeStamp", required=false)
+        @field:Attribute(name = "TimeStamp", required = false)
         var timeStamp: Date? = null
-        @field:Attribute(name = "TrackingReference", required=false)
+        @field:Attribute(name = "TrackingReference", required = false)
         var trackingReference: String? = null
     }
 
     class POSdata {
         @field:Element(name = "POSTimeStamp", required = true)
         lateinit var posTimeStamp: Date
-        @field:Element(name = "ServiceLevel", required=false)
+        @field:Element(name = "ServiceLevel", required = false)
         var serviceLevel: String? = null
-        @field:Element(name = "ShiftNumber", required=false)
-        var shiftNumber: BigInteger? = null
-        @field:Element(name = "ClerkID", required=false)
+        @field:Element(name = "ShiftNumber", required = false)
+        var shiftNumber: Int? = null
+        @field:Element(name = "ClerkID", required = false)
         var clerkID: Int? = null
-        @field:Element(name = "OutdoorPosition", required=false)
-        var outdoorPosition: BigInteger? = null
-        @field:Element(name = "ManualPAN", required=false)
+        @field:Element(name = "OutdoorPosition", required = false)
+        var outdoorPosition: Int? = null
+        @field:Element(name = "ManualPAN", required = false)
         var manualPAN: Boolean? = null
-        @field:Element(name = "Contactless", required=false)
+        @field:Element(name = "Contactless", required = false)
         var contactless: Boolean? = null
-        @field:Element(name = "ContactlessReceipt", required=false)
+        @field:Element(name = "ContactlessReceipt", required = false)
         var contactlessReceipt: Boolean? = null
-        @field:Element(name = "Chip", required=false)
+        @field:Element(name = "Chip", required = false)
         var chip: Boolean? = null
-        @field:Element(name = "Swipe", required=false)
+        @field:Element(name = "Swipe", required = false)
         var swipe: Boolean? = null
-        @field:Element(name = "ClerkPermission", required=false)
+        @field:Element(name = "ClerkPermission", required = false)
         var clerkPermission: String? = null
-        @field:Element(name = "TightControl", required=false)
+        @field:Element(name = "TightControl", required = false)
         var tightControl: Boolean? = null
-        @field:Element(name = "SplitPayment", required=false)
+        @field:Element(name = "SplitPayment", required = false)
         var splitPayment: Boolean? = null
-        @field:Element(name = "VoiceReferral", required=false)
+        @field:Element(name = "VoiceReferral", required = false)
         var voiceReferral: Boolean? = null
-        @field:Element(name = "TransactionNumber", required=false)
-        var transactionNumber: BigInteger? = null
-        @field:Element(name = "Reference", required=false)
+        @field:Element(name = "TransactionNumber", required = false)
+        var transactionNumber: Int? = null
+        @field:Element(name = "Reference", required = false)
         var reference: String? = null
-        @field:Element(name = "CardHolderLocation", required=false)
+        @field:Element(name = "CardHolderLocation", required = false)
         var cardHolderLocation: String? = null
-        @field:Attribute(name = "LanguageCode", required=false)
+        @field:Attribute(name = "LanguageCode", required = false)
         var languageCode: LanguageCodeType? = null
     }
 }
